strategy pattern for merging
